% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{prop}
\alias{prop}
\title{Get the proportion of each number in an array}
\usage{
prop(x)
}
\arguments{
\item{x}{an array of numbers}
}
\value{
an array of proportions
}
\description{
Get the proportion of each number in an array
}
\examples{
prop(c(1,2,3,4,5))
}
